Added new properties ProgId, ExternalLink and Image to the ExcelOleObject class.
Changed the type of OleObject in ExcelDrawingAsType.
