+++ src/EPPlus/Drawing/ExcelImage.cs
@@ -25,22 +25,15 @@ namespace OfficeOpenXml.Drawing
     /// <summary>
     /// Represents an image 
     /// </summary>
-    public class ExcelImage
+    public class ExcelImage : ExcelImageReadOnly
     {
-        IPictureContainer _container;
-        ePictureType[] _restrictedTypes = new ePictureType[0];
-        internal ExcelImage(IPictureContainer container, ePictureType[] restrictedTypes=null)
+        internal ExcelImage(IPictureContainer container, ePictureType[] restrictedTypes = null) : base(container, restrictedTypes) 
         {
-            _container = container;
-            if (restrictedTypes != null)
-            {
-                _restrictedTypes = restrictedTypes;
-            }
         }
         /// <summary>
         /// Creates an ExcelImage to be used as template for adding images.
         /// </summary>
-        public ExcelImage()
+        public ExcelImage() 
         {
 
         }
@@ -48,7 +41,7 @@ namespace OfficeOpenXml.Drawing
         /// Creates an ExcelImage to be used as template for adding images.
         /// </summary>
         /// <param name="imagePath">A path to the image file to load</param>
-        public ExcelImage(string imagePath)
+        public ExcelImage(string imagePath) 
         {
             SetImage(imagePath);
         }
@@ -78,41 +71,6 @@ namespace OfficeOpenXml.Drawing
         {
             SetImage(imageBytes, pictureType);
         }
-        /// <summary>
-        /// If this object contains an image.
-        /// </summary>
-        public bool HasImage
-        {
-            get
-            {
-                return Type.HasValue;
-            }
-        }
-        /// <summary>
-        /// The type of image.
-        /// </summary>
-        public ePictureType? Type
-        {
-            get;
-            internal set;
-        }
-
-        /// <summary>
-        /// The image as a byte array.
-        /// </summary>
-        public byte[] ImageBytes 
-        { 
-            get;
-            internal set; 
-        }
-        /// <summary>
-        /// The image bounds and resolution
-        /// </summary>
-        public ExcelImageInfo Bounds
-        {
-            get;            
-            internal set;
-        } = new ExcelImageInfo();
         /// <summary>
         /// Sets a new image. 
         /// </summary>
@@ -156,7 +114,8 @@ namespace OfficeOpenXml.Drawing
         /// <param name="pictureType">The type of image.</param>
         public ExcelImage SetImage(byte[] imageBytes, ePictureType pictureType)
         {
-            return SetImage(imageBytes, pictureType, true);            
+            SetImage(imageBytes, pictureType, true);
+            return this;
         }
         /// <summary>
         /// Sets a new image. 
@@ -169,7 +128,8 @@ namespace OfficeOpenXml.Drawing
             {
                 throw new ArgumentNullException("Image type must not be null");
             }
-            return SetImage(image.ImageBytes, image.Type.Value, true);
+            SetImage(image.ImageBytes, image.Type.Value, true);
+            return this;
         }
 
         /// <summary>
@@ -181,7 +141,7 @@ namespace OfficeOpenXml.Drawing
         {
             if(imageStream is MemoryStream ms)
             {
-                return SetImage(ms.ToArray(), pictureType, true);
+                SetImage(ms.ToArray(), pictureType, true);
             }
             else
             {
@@ -193,8 +153,9 @@ namespace OfficeOpenXml.Drawing
                 imageStream.Seek(0, SeekOrigin.Begin);
                 var r = imageStream.Read(byRet, 0, (int)imageStream.Length);
 
-                return SetImage(byRet, pictureType);
+                SetImage(byRet, pictureType);
             }
+            return this;
         }
 #if !NET35
         /// <summary>
@@ -206,7 +167,7 @@ namespace OfficeOpenXml.Drawing
         {
             if (imageStream is MemoryStream ms)
             {
-                return SetImage(ms.ToArray(), pictureType, true);
+                SetImage(ms.ToArray(), pictureType, true);
             }
             else
             {
@@ -218,8 +179,9 @@ namespace OfficeOpenXml.Drawing
                 imageStream.Seek(0, SeekOrigin.Begin);
                 var r = await imageStream.ReadAsync(byRet, 0, (int)imageStream.Length);
 
-                return SetImage(byRet, pictureType);
+                SetImage(byRet, pictureType);
             }
+            return this;
         }
         /// <summary>
         /// Sets a new image. 
@@ -253,139 +215,11 @@ namespace OfficeOpenXml.Drawing
             var fs = imageFile.OpenRead();
             var b = new byte[fs.Length];
             var r = await fs.ReadAsync(b, 0, b.Length);
-            return SetImage(b, type, true);
+            SetImage(b, type, true);
+            return this;
         }
 
 #endif
-        internal ExcelImage SetImageNoContainer(byte[] image, ePictureType pictureType)
-        {
-            Type = pictureType;
-            if (pictureType == ePictureType.Wmz ||
-               pictureType == ePictureType.Emz)
-            {
-                var img = ImageReader.ExtractImage(image, out ePictureType? pt);
-                if (pt.HasValue)
-                {
-                    throw new ArgumentException($"Image is not of type {pictureType}.", nameof(image));
-                }
-                else
-                {
-                    ImageBytes = img;
-                    pictureType = pt.Value;
-                }
-            }
-            else
-            {
-                ImageBytes = image;
-            }
-            using(var ms = RecyclableMemory.GetStream(image))
-            {
-                var imageHandler = new GenericImageHandler();
-                if (imageHandler.GetImageBounds(ms, pictureType, out double width, out double height, out double horizontalResolution, out double verticalResolution))
-                {
-                    Bounds.Width = width;
-                    Bounds.Height = height;
-                    Bounds.HorizontalResolution = horizontalResolution;
-                    Bounds.VerticalResolution = verticalResolution;
-                }
-                else
-                {
-                    throw (new InvalidOperationException($"The image format is not supported: {pictureType} or the image is corrupt "));
-                }
-            }
-            return this;
-
-        }
-        internal ExcelImage SetImage(byte[] image, ePictureType pictureType, bool removePrevImage)
-        {
-            if(_container == null)
-            {
-                return SetImageNoContainer(image, pictureType);
-            }
-            else
-            {
-                return SetImageContainer(image, pictureType, removePrevImage);
-            }
-        }
-
-        private ExcelImage SetImageContainer(byte[] image, ePictureType pictureType, bool removePrevImage)
-        {
-            ValidatePictureType(pictureType);
-            Type = pictureType;
-            if (pictureType == ePictureType.Wmz ||
-               pictureType == ePictureType.Emz)
-            {
-                var img = ImageReader.ExtractImage(image, out ePictureType? pt);
-                if (pt.HasValue)
-                {
-                    throw new ArgumentException($"Image is not of type {pictureType}.", nameof(image));
-                }
-                else
-                {
-                    if (string.IsNullOrEmpty(_container.ImageHash) == false && removePrevImage)
-                    {
-                        RemoveImageContainer();
-                    }
-                    ImageBytes = img;
-                    pictureType = pt.Value;
-                }
-            }
-            else
-            {
-                if (removePrevImage && string.IsNullOrEmpty(_container.ImageHash) == false)
-                {
-                    RemoveImageContainer();
-                }
-                ImageBytes = image;
-            }
-
-            PictureStore.SavePicture(image, _container, pictureType);
-
-            using (var ms = RecyclableMemory.GetStream(image))
-            {
-                if (_container.RelationDocument.Package.Settings.ImageSettings.GetImageBounds(ms, pictureType, out double width, out double height, out double horizontalResolution, out double verticalResolution))
-                {
-                    Bounds.Width = width;
-                    Bounds.Height = height;
-                    Bounds.HorizontalResolution = horizontalResolution;
-                    Bounds.VerticalResolution = verticalResolution;
-                }
-                else
-                {
-                    throw (new InvalidOperationException($"Image format not supported or: {pictureType} or corrupt image"));
-                }
-            }
-            _container.SetNewImage();
-            return this;
-        }
-
-        internal void SetRestrictedTypes(ePictureType[] restrictedTypes)
-        {
-            _restrictedTypes = restrictedTypes;
-        }
-
-        private void ValidatePictureType(ePictureType pictureType)
-        {
-            if (Array.Exists(_restrictedTypes, x => x == pictureType))
-            {
-                throw new InvalidOperationException($"Picture type {pictureType} is not supported for this operation.");
-            }
-        }
-
-        internal void RemoveImage()
-        {
-            RemoveImageContainer();
-            ImageBytes = null;
-            Type = null;
-            Bounds = new ExcelImageInfo();
-        }
-        private void RemoveImageContainer()
-        {
-            _container.RemoveImage();
-            _container.RelPic = null;
-            _container.ImageHash = null;
-            _container.UriPic = null;
-        }
     }
 }
